Flesh out form test + refactor helper
